fixe mira store
